requirements.txt
streamlit>=1.37.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
scikit-learn>=1.2.0
kaggle>=1.5.0
